requirements.txt
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
requests>=2.31.0
websockets>=12.0
streamlit>=1.30.0
plotly>=5.18.0
pydantic>=2.5.0
python-dotenv>=1.0.0
